requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
shap==0.44.0
mysql-connector-python==8.2.0
pymysql==1.1.0
pydantic==2.5.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-json-logger==2.0.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0 